requirements.txt
Django==1.4.2
MySQL-python==1.2.3
pillow==2.1.0
South==0.7.6
amqp==1.0.11
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.26
celery==3.0.17
crypto==1.0.0
distribute==0.6.24
django-celery==3.0.17
django-extensions==1.1.1
django-indexer==0.3.0
django-notification==0.2
django-paging==0.2.4
django-registration==0.8
django-reversion==1.7
django-rosetta==0.7.1
django-sentry==1.13.5
django-smart-selects==1.0.6
django-swingtime==0.2.2
django-templatetag-sugar==0.1
eventlet==0.12.1
greenlet==0.4.0
html5lib==0.95
kombu==2.5.10
lockfile==0.9.1
pisa==3.0.33
pyasn1==0.1.6
pycrypto==2.6
http://slowchop.com/pygooglechart/download/pygooglechart-0.3.0.tar.gz
python-daemon==1.6
python-dateutil==1.5
pytz==2013b
raven==3.3.3
reportlab==2.7
rsa==3.1.1
six==1.3.0
wsgiref==0.1.2
